Refactor out code for ADBench evals in C++ tools
