+++ tools/finite/Dockerfile
@@ -11,12 +11,9 @@ COPY python /home/gradbench/python
 ENV PYTHONPATH=/home/gradbench/python/gradbench
 
 COPY cpp /home/gradbench/cpp
-
 RUN make -C cpp
 
 COPY tools/finite/ /home/gradbench/tools/finite
-
-RUN make -j -B -C tools/finite
 
 ENTRYPOINT ["python3", "/home/gradbench/tools/finite/run.py"]
 LABEL org.opencontainers.image.source=https://github.com/gradbench/gradbench
