+++ tools/tapenade/gmm.py
@@ -1,27 +1,3 @@
-import json
-import subprocess
-import tempfile
-from os import listdir
+from gradbench.cpp_main import functions
 
-
-def compile():
-    # Nothing to do here. We assume everything is precompiled.
-    return True
-
-
-def objective(input):
-    with tempfile.NamedTemporaryFile("w") as tmp:
-        json.dump(input, tmp)
-        tmp.flush()
-        return subprocess.run(
-            ["tools/tapenade/run_gmm", tmp.name, "F"], text=True, capture_output=True
-        )
-
-
-def jacobian(input):
-    with tempfile.NamedTemporaryFile("w") as tmp:
-        json.dump(input, tmp)
-        tmp.flush()
-        return subprocess.run(
-            ["tools/tapenade/run_gmm", tmp.name, "J"], text=True, capture_output=True
-        )
+globals().update(functions(__file__))
